+++ plugin/src/main/scala/com/stratio/cassandra/lucene/partitioning/PartitionerOnToken.scala
@@ -17,33 +17,36 @@ package com.stratio.cassandra.lucene.partitioning
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.stratio.cassandra.lucene.IndexException
+import org.apache.cassandra.config.CFMetaData
 import org.apache.cassandra.db._
 import org.apache.cassandra.dht.Token
 
-/** [[Partitioner]] based on the partition key token.
+/** [[Partitioner]] based on the partition key token. Rows will be stored in an index partition
+  * determined by the hash of the partition key token. Partition-directed searches will be routed to
+  * a single partition, increasing performance. However, token range searches will be routed to all
+  * the partitions, with a slightly lower performance.
   *
-  * Partitioning on token guarantees a good load balancing between partitions while speeding up
-  * partition-directed searches to the detriment of token range searches.
+  * This partitioner guarantees an excellent load balancing between index partitions.
   *
-  * @param partitions the number of partitions
+  * @param partitions the number of index partitions per node
   * @author Andres de la Pena `adelapena@stratio.com`
   */
-case class PartitionerOnToken(@JsonProperty("partitions") partitions: Int) extends Partitioner {
+case class PartitionerOnToken(partitions: Int) extends Partitioner {
 
   if (partitions <= 0) throw new IndexException(
     s"The number of partitions should be strictly positive but found $partitions")
 
-  /** @inheritdoc */
+  /** @inheritdoc*/
   private[this] def partition(token: Token): Int =
     (Math.abs(token.getTokenValue.asInstanceOf[Long]) % partitions).toInt
 
-  /** @inheritdoc */
+  /** @inheritdoc*/
   override def numPartitions: Int = partitions
 
-  /** @inheritdoc */
+  /** @inheritdoc*/
   override def partition(key: DecoratedKey): Int = partition(key.getToken)
 
-  /** @inheritdoc */
+  /** @inheritdoc*/
   override def partitions(command: ReadCommand): List[Int] = command match {
     case c: SinglePartitionReadCommand => List(partition(c.partitionKey))
     case c: PartitionRangeReadCommand =>
@@ -52,6 +55,19 @@ case class PartitionerOnToken(@JsonProperty("partitions") partitions: Int) exten
       val stop = range.stopKey().getToken
       if (start.equals(stop) && !start.isMinimum) List(partition(start)) else allPartitions
     case _ => throw new IndexException(s"Unsupported read command type: ${command.getClass}")
+  }
+
+}
+
+/** Companion object for [[PartitionerOnToken]]. */
+object PartitionerOnToken {
+
+  /** [[PartitionerOnToken]] builder.
+    *
+    * @param partitions the number of index partitions per node
+    */
+  case class Builder(@JsonProperty("partitions") partitions: Int) extends Partitioner.Builder {
+    override def build(metadata: CFMetaData): PartitionerOnToken = PartitionerOnToken(partitions)
   }
 
 }